Refactored GuideView
